requirements.txt
discord.py==1.5.1
asyncio
git+https://github.com/Rapptz/discord-ext-menus
pendulum
googletrans
asyncpg
apscheduler
psutil
git+https://github.com/Gorialis/jishaku@master#egg=jishaku
motor
pymongo[srv]
DiscordUtils